Update printer.cfg
Changes for the bigger bed of the Plus compared to the Pro and 6 instead 4 screws for the print bed added in the config.

safe_z_hame, bed_mesh and sscrews_tilt_adjust edited. 
screws_tilt_adjust also moved below bed_mesh because I think it makes more sense to have it before the gcode_macros at the bottom.

I think it would be a good idea to move all gcode_macros to the bottom of the file to have them separated from the printer config but I didn't want to change the whole file.

Also added a comment for [exclude_object]
It must be disabled for Sonic Pad.